Reformatted with astyle -z2 -r -n --style=google -T -Z ./*.c

+++ others/quipu/uips/dish/pipe.c
@@ -4,7 +4,7 @@
 static char *rcsid = "$Header: /xtel/isode/isode/others/quipu/uips/dish/RCS/pipe.c,v 9.0 1992/06/16 12:44:21 isode Rel $";
 #endif
 
-/* 
+/*
  * $Header: /xtel/isode/isode/others/quipu/uips/dish/RCS/pipe.c,v 9.0 1992/06/16 12:44:21 isode Rel $
  *
  *
@@ -47,7 +47,7 @@ char	*argv[];
 	char			buffer [BUFSIZ];
 	char			dishname [BUFSIZ];
 	char 			* ptr;
-	
+
 	isodetailor (argv[0], 1);
 
 	if ((sd = start_tcp_client ((struct sockaddr_in *) 0, 0)) == NOTOK) {
@@ -59,25 +59,26 @@ char	*argv[];
 		exit (-21);
 
 	if (join_tcp_server (sd, sin) == NOTOK) {
-	        int	pid;
+		int	pid;
 		(void) close_tcp_socket (sd);
 
-fork_again: ;
+fork_again:
+		;
 		switch (pid = vfork ()) {
 		case 0:
 			/* child */
 			(void) close_tcp_socket (sd);
 			(void) strcpy (dishname,
-				       _isodefile (isodebinpath, "dish"));
+						   _isodefile (isodebinpath, "dish"));
 			{
 				int i, nfds = getdtablesize ();
 
 				for (i = fileno(stderr) + 1; i < nfds; i++)
 					(void) close (i);
 			}
- 			(void) execl (dishname, "dish","-pipe",NULLCP);
- 			(void) fprintf (stderr, "unable to exec ");
- 			perror (dishname);
+			(void) execl (dishname, "dish","-pipe",NULLCP);
+			(void) fprintf (stderr, "unable to exec ");
+			perror (dishname);
 			_exit (-21);
 		case -1:
 			perror ("fork");
@@ -86,23 +87,23 @@ fork_again: ;
 		default:
 			/* parent */
 			for (;;) {
-			    if ((sd = start_tcp_client ((struct sockaddr_in *) 0,
-							0)) == NOTOK) {
-				perror("start_tcp_client");
-				exit(-23);
-			    }
-			    if (join_tcp_server (sd, sin) != NOTOK)
-				break;
+				if ((sd = start_tcp_client ((struct sockaddr_in *) 0,
+											0)) == NOTOK) {
+					perror("start_tcp_client");
+					exit(-23);
+				}
+				if (join_tcp_server (sd, sin) != NOTOK)
+					break;
 
-			    /* need to introduce a timeout !!! */
-			    (void) close_tcp_socket (sd);
+				/* need to introduce a timeout !!! */
+				(void) close_tcp_socket (sd);
 
-			    sleep (5);
+				sleep (5);
 
-			    if (kill (pid, 0) == NOTOK) {
-				(void) fprintf (stderr,"Trying again...\n");
-				goto fork_again;
-			    }
+				if (kill (pid, 0) == NOTOK) {
+					(void) fprintf (stderr,"Trying again...\n");
+					goto fork_again;
+				}
 			}
 			break;
 		}
@@ -130,7 +131,8 @@ fork_again: ;
 
 	for (;;) {
 		if ((res = recv(sd, buffer, BUFSIZ - 1, 0)) == -1) {
-err_recv: ;
+err_recv:
+			;
 			perror ("recv");
 			(void) close_tcp_socket (sd);
 			exit (-26);
@@ -147,49 +149,49 @@ err_recv: ;
 			while ( (res = recv(sd, buffer, BUFSIZ, 0)) > 0)
 				(void) write (2,buffer,res);
 			(void) close_tcp_socket (sd);
-			exit (1);			
+			exit (1);
 		} else if ((*buffer == '1') || (*buffer == '3')) {
 			int eval;
 			eval = (*buffer == '1' ? 0 : 2);
 			if (res > 1)
-	                        (void) write (1,&buffer[1],--res);
-                        while ( (res = recv(sd, buffer, BUFSIZ, 0)) > 0)
-                                (void) write (1,buffer,res);
-                        (void) close_tcp_socket (sd);
-                        exit (eval);
-                } else {		/* 'e', 'y', 'm', or 'p' */
-		        register char  *cp, *ep;
+				(void) write (1,&buffer[1],--res);
+			while ( (res = recv(sd, buffer, BUFSIZ, 0)) > 0)
+				(void) write (1,buffer,res);
+			(void) close_tcp_socket (sd);
+			exit (eval);
+		} else {		/* 'e', 'y', 'm', or 'p' */
+			register char  *cp, *ep;
 			char            where[BUFSIZ];
 
 			cp = buffer + res - 1;
 			ep = buffer + sizeof buffer - 1;
 			while (*cp != '\n') {
-			    ++cp;
-			    switch (res = recv (sd, cp, ep - cp, 0)) {
+				++cp;
+				switch (res = recv (sd, cp, ep - cp, 0)) {
 				case NOTOK:
-				    goto err_recv;
+					goto err_recv;
 				case OK:
-				    (void) fprintf (stderr,
-					     "eof reading '%c' directive\n",
-					     *buffer);
-				    exit (-28);
+					(void) fprintf (stderr,
+									"eof reading '%c' directive\n",
+									*buffer);
+					exit (-28);
 				default:
-				    cp += res - 1;
-				    if (cp < ep)
-					continue;
-				    (void) fprintf (stderr,
-					     "'%c' directive exceeds %d octets\n",
-					     *buffer, sizeof buffer - 1);
-				    exit(-29);
-			    }
+					cp += res - 1;
+					if (cp < ep)
+						continue;
+					(void) fprintf (stderr,
+									"'%c' directive exceeds %d octets\n",
+									*buffer, sizeof buffer - 1);
+					exit(-29);
+				}
 			}
 			*cp = NULL;
 
 			if (*buffer == 'e') {
 				if (system (&buffer[1]))
-				    (void) strcpy (where, "e");
+					(void) strcpy (where, "e");
 				else
-				    (void) getcwd (where, sizeof where);
+					(void) getcwd (where, sizeof where);
 			} else if (*buffer == 'm') {
 				(void) fprintf (stderr, "\n%s\n", buffer + 1);
 				(void) strcpy (where, "m");
@@ -197,13 +199,13 @@ err_recv: ;
 				(void) fprintf (stderr,"%s",buffer + 1);
 				(void) fgets (where, sizeof where, stdin);
 				if (cp = index (where, '\n'))
-				    *cp = NULL;
+					*cp = NULL;
 			} else {	/* 'p' */
 				(void) sprintf (where,
-					        "Enter password for \"%s\": ",
-						buffer + 1);
+								"Enter password for \"%s\": ",
+								buffer + 1);
 				(void) sprintf (where, "p%s",
-						getpassword (where));
+								getpassword (where));
 			}
 			(void) strcat (where, "\n");
 
@@ -246,11 +248,10 @@ char          **argv;
 	(void) umask (0);
 	(void) sprintf (retfile, "/tmp/dish%d", getpid ());
 	if ((ptr = getenv ("DISHPROC")) == NULLCP) {
-	    (void) sprintf (sendfile, "/tmp/dish-%d", getppid ());
-	    (void) setenv ("DISHPROC", sendfile);
-	}
-	else
-	    (void) strcpy (sendfile, ptr);
+		(void) sprintf (sendfile, "/tmp/dish-%d", getppid ());
+		(void) setenv ("DISHPROC", sendfile);
+	} else
+		(void) strcpy (sendfile, ptr);
 
 	setbuf (stdout,NULLCP);
 	setbuf (stderr,NULLCP);
@@ -259,7 +260,7 @@ char          **argv;
 		(void) fprintf (stderr,"Can't create pipe '%s'\n",retfile);
 		exit (-12);
 	}
-	
+
 	for (i = 1; i <= 15; i++)
 		(void) signal (i, pipe_quit);
 
@@ -334,71 +335,71 @@ char          **argv;
 			(void) close (fd);
 			(void) close (wfd);
 			(void) unlink (retfile);
-			exit (-1);			
+			exit (-1);
 		} else if ((*buffer == '1') || (*buffer == '3')) {
 			int eval;
 			eval = (*buffer == '1' ? 0 : 2);
-                        (void) write (1,&buffer[1],--res);
-                        while ( (res = read (fd, buffer, BUFSIZ)) > 0)
-                                (void) write (1,buffer,res);
+			(void) write (1,&buffer[1],--res);
+			while ( (res = read (fd, buffer, BUFSIZ)) > 0)
+				(void) write (1,buffer,res);
 			(void) close (fd);
 			(void) close (wfd);
 			(void) unlink (retfile);
-                        exit (eval);
-		else {		/* 'e', 'y', 'm', or 'p' */
-			int             res2,
-			                nfd;
-			char            where[BUFSIZ];
+			exit (eval);
+			else {		/* 'e', 'y', 'm', or 'p' */
+				int             res2,
+								nfd;
+				char            where[BUFSIZ];
 
-			if (*buffer == 'e') {
-				if (system (&buffer[1]))
-				    (void) strcpy (where, "e");
-				else
-				    (void) getcwd (where, sizeof where);
-			} else if (*buffer == 'm') {
-				(void) fprintf (stderr, "\n%s\n", buffer + 1);
-				(void) strcpy (where, "m");
-			} else if (*buffer == 'y') {
-			        char   *cp;
+				if (*buffer == 'e') {
+					if (system (&buffer[1]))
+						(void) strcpy (where, "e");
+					else
+						(void) getcwd (where, sizeof where);
+				} else if (*buffer == 'm') {
+					(void) fprintf (stderr, "\n%s\n", buffer + 1);
+					(void) strcpy (where, "m");
+				} else if (*buffer == 'y') {
+					char   *cp;
 
-				(void) fprintf (stderr,"%s",buffer + 1);
-				(void) fgets (where, sizeof where, stdin);
-				if (cp = index (where, '\n'))
-				    *cp = NULL;
-			} else {	/* 'p' */
-				(void) sprintf (where,
-					        "Enter password for \"%s\": ",
-						buffer + 1);
-				(void) sprintf (where, "p%s",
-						getpassword (where));
-			}
-			(void) strcat (where, "\n");
+					(void) fprintf (stderr,"%s",buffer + 1);
+					(void) fgets (where, sizeof where, stdin);
+					if (cp = index (where, '\n'))
+						*cp = NULL;
+				} else {	/* 'p' */
+					(void) sprintf (where,
+									"Enter password for \"%s\": ",
+									buffer + 1);
+					(void) sprintf (where, "p%s",
+									getpassword (where));
+				}
+				(void) strcat (where, "\n");
 
-			if ((nfd = open (sendfile, O_WRONLY)) == -1) {
-				(void) fprintf (stderr, "re-write open error\n");
-				(void) close (wfd);
-				(void) close (fd);
-				(void) unlink (retfile);
-				exit (-9);
-			}
-			if ((res2 = write (nfd, where, strlen (where))) == -1) {
-				(void) fprintf (stderr, "Write to DUA failed (%d)... Please retry\n", res2);
-				(void) close (fd);
+				if ((nfd = open (sendfile, O_WRONLY)) == -1) {
+					(void) fprintf (stderr, "re-write open error\n");
+					(void) close (wfd);
+					(void) close (fd);
+					(void) unlink (retfile);
+					exit (-9);
+				}
+				if ((res2 = write (nfd, where, strlen (where))) == -1) {
+					(void) fprintf (stderr, "Write to DUA failed (%d)... Please retry\n", res2);
+					(void) close (fd);
+					(void) close (nfd);
+					(void) close (wfd);
+					(void) unlink (retfile);
+					exit (-10);
+				}
 				(void) close (nfd);
-				(void) close (wfd);
-				(void) unlink (retfile);
-				exit (-10);
 			}
-			(void) close (nfd);
 		}
 	}
-}
 
-void pipe_quit (sig)
-int     sig;
-{
-	(void) unlink (retfile);
-	(void) fprintf (stderr,"(signal %d) exiting...\n",sig);
-	exit(0);
-}
+	void pipe_quit (sig)
+	int     sig;
+	{
+		(void) unlink (retfile);
+		(void) fprintf (stderr,"(signal %d) exiting...\n",sig);
+		exit(0);
+	}
 #endif
